+++ src/main/java/com/springVaadin/test/repos/CustomerRepository.java
@@ -3,12 +3,11 @@ package com.springVaadin.test.repos;
 import com.springVaadin.test.domains.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, BigDecimal> {
+public interface CustomerRepository extends JpaRepository<Customer, BigInteger> {
 
     List<Customer> findByLastNameStartsWithIgnoreCase(String lastName);
